+++ src/Akeneo/ReferenceEntity/tests/back/Integration/Search/Elasticsearch/Record/FindIdentifiersForQueryTest.php
@@ -497,6 +497,7 @@ class FindIdentifiersForQueryTest extends SearchIntegrationTestCase
             'identifiers' => ['brand_bangolufsen', 'brand_kartell'],
             'matches_count' => 3
         ], $matchingIdentifiers->normalize());
+        self::assertSame('kartell', $matchingIdentifiers->lastSortValue);
     }
 
     /**