+++ CHANGELOG.md
@@ -75,6 +75,7 @@
 - PIM-10021: Product link rule too long to be completed
 - PIM-10026: Avoid session persistance for API
 - PIM-10028: Fix the ETA time displayed when executing a rule with several actions
+- PIM-10042: Fix infinite loop when using get all reference entity records API endpoint
 
 ## Improvements